refactor(tasks-gate): split task-scope.js into focused modules
Split monolithic task-scope.js (781 lines) into three focused modules to satisfy
lint complexity and max-lines rules while preserving public API:
- task-scope-globs.js: glob matching logic
- task-scope-validators.js: validation helpers
- task-scope-test-validator.js: test-specific validation

Fixes GH-516.
